feat(subscription): let a meter count in fractions, opt-in per meter
Some plans sell a unit that genuinely has halves — storage in GB, time in
hours — and every metered quantity in the module was a long.

Quantities become decimal, stored as BSON Decimal128, from the plan's
allowance through the ledger and the atomic counter to the current-usage
projection. Exact base ten rather than a double, because a reversal has to
cancel the entry it compensates to the last place; a binary residue would
sit in the balance for the rest of the period and be billed.

Fractions are opt-in per meter. PlanMeter.QuantityScale is how many decimal
places a meter accepts, zero by default and at most six, and it governs that
meter's allowance, its carry-forward cap, its rate-band bounds, a trial
grant for it, and every quantity recorded against it. Opt-in because the two
kinds live on the same plan: a storage meter holds 512.5, a screening meter
has no half. Before this, { "quantity": 0.5 } was refused for free — JSON
cannot bind a fraction to a long — and widening the type globally would have
spent that guard silently, making a stray 1.3333 from a calling product
recordable and billable. A plan already stored has no such field, reads as
zero, and cannot behave differently from before.

No data migration. The driver's default decimal serializer reads Int64 and
Int32 back as decimals, and $inc with a Decimal128 promotes the field in
place, so every counter and ledger row already written loads unchanged. That
mattered because the alternative — fixed-point scaled integers — would have
required rewriting the append-only usage ledger, which is the authority every
past invoice was computed from. Both halves are pinned: the serializer in
SubscriptionEntitySerializationTests, the write in
FractionalUsageQuantityIntegrationTests against a real mongod.

Rating stays exact through the tier walk and rounds once per meter, half away
from zero. Rounding once is what keeps re-banding a rate table without
changing any of its prices from changing the bill, and keeps a band breakdown
adding up to the figure that was rounded. Bands become closed above and open
below so a fraction lands in exactly one, and a band reports its first
quantity as one step above its open bound — which at scale zero is the whole
unit it has always started at, so a whole-unit meter still reports 1-400 and
401 onwards. The preview needed no rounding rule of its own: it already
defines the additional charge as the difference of two rated totals, so
parity survives by construction.

Also repairs an unrelated pre-existing defect found in the initializers the
scale had to be added to: CarryForwardCap was never copied from the request
to the stored plan, nor from the plan into the subscription snapshot. It was
validated as mandatory, reported as null by three responses, and read as "no
cap" by MeterAllowance.CarriedIn — so a dormant subscription banked allowance
forever, the exact outcome requiring the cap is meant to prevent. Invisible
to the suite because every test set the field directly on a snapshot.

Breaking for any caller that meets a meter with a scale above zero: included,
used, remaining and overage on UsageResponse, and the plan and preview
quantity fields, may now be non-integral. There is no way to version a
number's domain.

+++ server/Subscription.DomainService/Entities/PlanEntitlement.cs
@@ -21,7 +21,7 @@ public sealed class PlanEntitlement
         EntitlementLimitKind.Boolean;
 
     /// <summary>The cap, when the kind is a count.</summary>
-    public long? Limit { get; set; }
+    public decimal? Limit { get; set; }
 
     /// <summary>The meter that draws this entitlement down, when the kind is a count.</summary>
     public string? MeterKey { get; set; }
